# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dokumentasi API",
        "description": "Lorem Ipsum",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/hr/employees": {
            "get": {
                "tags": [
                    "HR - Employees"
                ],
                "summary": "Mendapatkan daftar karyawan",
                "description": "List semua karyawan aktif. HR bisa melihat dan mengelola semua.",
                "operationId": "386d776fa54946fe11b705377c215eed",
                "parameters": [
                    {
                        "name": "department_id",
                        "in": "query",
                        "description": "Filter berdasarkan departemen",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "position_id",
                        "in": "query",
                        "description": "Filter berdasarkan posisi",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "contract_status",
                        "in": "query",
                        "description": "Filter berdasarkan status kontrak",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Pencarian berdasarkan nama atau email",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Jumlah data per halaman",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Daftar karyawan berhasil diambil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/Employee"
                                                    }
                                                },
                                                "total": {
                                                    "type": "integer",
                                                    "example": 50
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/hr/employees/{id}": {
            "get": {
                "tags": [
                    "HR - Employees"
                ],
                "summary": "Melihat detail karyawan",
                "description": "Detail data kepegawaian karyawan beserta atasan dan bawahan.",
                "operationId": "927c57094d04243a410c2cd6c9db3af7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID karyawan (employee.id)",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail karyawan berhasil diambil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Employee"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Karyawan tidak ditemukan"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "HR - Employees"
                ],
                "summary": "Update data karyawan",
                "description": "Update data kepegawaian karyawan (posisi, departemen, kontrak, supervisor).",
                "operationId": "c338c7d337faa31b35d25bc831af6f5b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID karyawan",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "department_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "position_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "supervisor_id": {
                                        "type": "integer",
                                        "example": 3
                                    },
                                    "join_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2023-05-01"
                                    },
                                    "contract_status": {
                                        "type": "string",
                                        "example": "permanent",
                                        "enum": [
                                            "permanent",
                                            "contract",
                                            "probation"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Data karyawan berhasil diupdate",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Data kepegawaian berhasil diupdate."
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Employee"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Karyawan tidak ditemukan"
                    },
                    "422": {
                        "description": "Validasi gagal (misal supervisor adalah diri sendiri)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server Error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Employee": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 3
                    },
                    "department_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "position_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "supervisor_id": {
                        "type": "integer",
                        "example": null,
                        "nullable": true
                    },
                    "join_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2023-01-01"
                    },
                    "contract_status": {
                        "type": "string",
                        "example": "permanent",
                        "enum": [
                            "permanent",
                            "contract",
                            "probation"
                        ]
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "ValidationError": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "The given data was invalid."
                    },
                    "errors": {
                        "type": "object",
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                },
                "type": "object"
            }
        }
    },
...
}